Test shinyjster app on all browsers

This method will test your shiny application using the shinyjster code you have provided on all of the available browsers shinyjster can test with on the given platform. For each browser, each app will be tested using run_jster_apps().

test_jster(
  apps = "../",
  browsers = c(selenium_chrome(), selenium_firefox(), if (platform() == "win" ||
    platform() == "mac") c(selenium_edge()), if (platform() == "win") c(selenium_ie()),
    if (platform() == "mac") c(selenium_safari())),
  type = c("serial", "lapply"),
  assert = TRUE,
  host = "127.0.0.1",
  port = NULL
)

Arguments

apps
Defaults the app in the directory above

browsers
By default, as many browsers as selenium support on the given platform

type
Single value to determine how applications are executed.

'serial'

Runs apps one after another using lapply. port will be random for each app unless specified.

'callr'

Runs apps using callr::r_bg using cores cores. port will be random for each app to allow concurrent execution.

'lapply'

Runs apps in succession using lapply. port will be random for each app unless specified.

assert
A logical value that determines if assert_jster() should be called on the return value

host
The IPv4 address that the application should listen on. Defaults to the shiny.host option, if set, or "127.0.0.1" if not. See Details.

port
The TCP port that the application should listen on. If the port is not specified, and the shiny.port option is set (with options(shiny.port = XX)), then that port will be used. Otherwise, use a random port between 3000:8000, excluding ports that are blocked by Google Chrome for being considered unsafe: 3659, 4045, 5060, 5061, 6000, 6566, 6665:6669 and 6697. Up to twenty random ports will be tried.

Value

A data frame with the columns appDir, successful, returnValue, and browser. One row of information per browser and apps combination.

See also